+++ packages/twenty-server/src/engine/core-modules/file/interfaces/file-folder.interface.ts
@@ -15,7 +15,10 @@ export const IMMUTABLE_FILE_CACHE_CONTROL = 'private, max-age=86400, immutable';
 
 export const PUBLIC_ASSET_CACHE_CONTROL = 'public, max-age=3600';
 
-export const PRESIGNED_URL_CACHE_CONTROL = 'private, no-store';
+// Responses embedding a short-lived presigned URL must never be stored:
+// a cached copy either 403s once the signature expires or leaks a live
+// presigned URL to clients that never authenticated.
+export const PRESIGNED_URL_NO_STORE_CACHE_CONTROL = 'private, no-store';
 
 export const fileFolderConfigs: Record<FileFolder, FileFolderConfig> = {
   [FileFolder.CorePicture]: {
