+++ packages/twenty-server/src/engine/metadata-modules/front-component/controllers/front-component.controller.ts
@@ -17,7 +17,7 @@ import {
   FileStorageException,
   FileStorageExceptionCode,
 } from 'src/engine/core-modules/file-storage/interfaces/file-storage-exception';
-import { PRESIGNED_URL_CACHE_CONTROL } from 'src/engine/core-modules/file/interfaces/file-folder.interface';
+import { PRESIGNED_URL_NO_STORE_CACHE_CONTROL } from 'src/engine/core-modules/file/interfaces/file-folder.interface';
 import { setFileResponseHeaders } from 'src/engine/core-modules/file/utils/set-file-response-headers.utils';
 
 import { WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.entity';
@@ -86,7 +86,7 @@ export class FrontComponentController {
       });
 
     if (fileResponse.type === 'redirect') {
-      res.setHeader('Cache-Control', PRESIGNED_URL_CACHE_CONTROL);
+      res.setHeader('Cache-Control', PRESIGNED_URL_NO_STORE_CACHE_CONTROL);
 
       return res.json({ url: fileResponse.presignedUrl });
     }
